requirements.txt
langchain-community
python-dotenv
groq
streamlit
requests
beautifulsoup4
langchain
langchain-groq
langchain-huggingface
sentence-transformers
selenium
bs4
pinecone
fastapi
uvicorn
